# swagger/openapi.yaml
openapi: 3.0.3
info:
  title: MCP Orchestrator API
  description: |
    LangChain agent orchestrator backed by [MCP-Router](http://localhost:8765) vector retrieval.

    **Flow:** `POST /chat` decomposes the user message into routing hops, calls the router's
    `POST /route` for each hop, binds returned tools for tool execution, and runs a GPT-4o agent.

    **Prerequisites:** `OPENAI_API_KEY` set for the orchestrator process; MCP-Router reachable
    at `router_url` (default `http://localhost:8765`).

    **Interactive docs:** FastAPI also exposes `/docs` and `/redoc` when the server is running.
  version: "0.1.0"
  contact:
    name: Research Proj

servers:
  # Relative to the URL where this file is served (e.g. …/swagger/openapi.yaml).
  # Avoids "Failed to fetch" when the UI is opened at http://127.0.0.1:8766 but the
  # spec hard-coded http://localhost:8766 (different origins → browser blocks).
  - url: /
    description: Same host as this page (recommended). Change `api_port` in config.yaml if not 8766.

tags:
  - name: admin
    description: Health and diagnostics
  - name: agent
    description: Chat and agent runs
  - name: servers
    description: MCP server registration and local execution registry
  - name: session
    description: Session / conversation state

paths:
  /health:
    get:
      tags: [admin]
      summary: Health check
      description: Returns orchestrator status, pings MCP-Router `/health`, and counts locally registered servers.
      operationId: health
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              examples:
                sample:
                  value:
                    status: ok
                    router:
                      status: ok
                      index_dir: "C:/path/to/MCP-Router/data/index"
                    registered_servers: 0

  /chat:
    post:
      tags: [agent]
      summary: Send a chat message
      description: |
        Runs the full pipeline: intent hops → router retrieval → LangChain agent (with MCP tools when available).
        Omit `session_id` to start a new conversation; reuse the returned `session_id` for multi-turn threads.
      operationId: chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
            examples:
              newSession:
                summary: New session
                value:
                  message: "Go to Bloomberg bcz I need to do some financial things. I neeed the historical data for Microsoft Stocks. Then go to Axiom and show me all the available data related to this. Then go to airtable to search for data on these stocks. ok now go to githb and push these data files to the finance branch."
              continueSession:
                summary: Continue session
                value:
                  message: "Now focus only on bugs."
                  session_id: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        "200":
          description: Agent completed (may answer without tools if router returns none).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "500":
          description: Internal error (e.g. LLM or router failure)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"

  /servers/register:
    post:
      tags: [servers]
      summary: Register MCP server (stdio)
      description: |
        Discovers tools from a local MCP server (stdio), registers them with MCP-Router (`POST /register/discover`),
        and stores connection details for orchestrator tool execution.
      operationId: registerServerStdio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterStdioRequest"
            example:
              name: "Filesystem"
              command: "npx"
              args: ["-y", "@modelcontextprotocol/server-filesystem", "C:/allowed/path"]
              description: "Read/write files under an allowed directory"
              discovery_timeout: 20.0
      responses:
        "200":
          description: Registered with router and local registry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "400":
          description: Discovery or router registration failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"

  /servers/register/sse:
    post:
      tags: [servers]
      summary: Register MCP server (HTTP / SSE-style JSON-RPC)
      description: |
        Discovers tools via HTTP JSON-RPC (`POST /register/discover/sse` on the router) and saves the URL for execution.
      operationId: registerServerSse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterSSERequest"
            example:
              name: "RemoteMCP"
              url: "http://127.0.0.1:9000/mcp"
              description: "Custom MCP over HTTP"
              discovery_timeout: 20.0
      responses:
        "200":
          description: Registered with router and local registry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "400":
          description: Discovery or router registration failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"

  /servers:
    get:
      tags: [servers]
      summary: List registered servers
      description: Returns the orchestrator's persisted server registry (connection info), not the router index dump.
      operationId: listServers
      responses:
        "200":
          description: Map of server name → registry entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerRegistryMap"

  /servers/{server_name}:
    delete:
      tags: [servers]
      summary: Remove server from local registry
      description: |
        Deletes the entry from `data/server_registry.json` only. The MCP-Router vector index is unchanged until rebuild/restart.
      operationId: removeServer
      parameters:
        - name: server_name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemoveServerResponse"
        "404":
          description: Server not found in local registry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"

  /session/clear:
    post:
      tags: [session]
      summary: Clear session
      description: Clears orchestrator in-memory chat history for the session and calls MCP-Router `POST /session/clear`.
      operationId: clearSession
      parameters:
        - name: session_id
          in: query
          required: true
          schema:
            type: string
          description: Session identifier returned by `POST /chat`
      responses:
        "200":
          description: Cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkStatus"

components:
  schemas:
    ChatRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: End-user prompt for this turn
        session_id:
          type: string
          nullable: true
          description: Existing session UUID; omit for a new thread

    RouterHop:
      type: object
      properties:
        server_intent:
          type: string
          description: 4–10 words, MCP domain (Policy.md)
        tool_intent:
          type: string
          description: Verb + object + constraints, 6–16 words (Policy.md)

    RouteResponseSummary:
      type: object
      description: Per-hop router metadata (subset of MCP-Router RouteResponse)
      properties:
        server_intent:
          type: string
        tool_intent:
          type: string
        confidence:
          type: string
          enum: [high, medium, low, disabled]
        recommended_k:
          type: integer
        null_route:
          type: boolean
        overlap_ambiguous:
          type: boolean
        tools_returned:
          type: array
          items:
            type: object
            properties:
              key:
                type: string
                example: "GitHub::list_pull_requests"
              name:
                type: string
              score:
                type: number
                format: float

    ChatResponse:
      type: object
      required: [session_id, answer, hops, tools_found, route_responses]
      properties:
        session_id:
          type: string
        answer:
          type: string
          description: Final assistant message for this turn
        hops:
          type: array
          items:
            $ref: "#/components/schemas/RouterHop"
        tools_found:
          type: array
          items:
            type: string
          description: LangChain tool names bound for this turn
        route_responses:
          type: array
          items:
            $ref: "#/components/schemas/RouteResponseSummary"

    RegisterStdioRequest:
      type: object
      required: [name, command]
      properties:
        name:
          type: string
          description: Unique server name (matches router registration)
        command:
          type: string
          description: Executable to spawn the MCP server (stdio)
        args:
          type: array
          items:
            type: string
          default: []
        description:
          type: string
          default: ""
        discovery_timeout:
          type: number
          format: float
          default: 20.0
          minimum: 2.0
          maximum: 120.0

    RegisterSSERequest:
      type: object
      required: [name, url]
      properties:
        name:
          type: string
        url:
          type: string
          format: uri
          description: MCP JSON-RPC HTTP endpoint
        description:
          type: string
          default: ""
        discovery_timeout:
          type: number
          format: float
          default: 20.0

    RegisterResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
        server_name:
          type: string
        transport:
          type: string
          enum: [stdio, sse]
        router:
          type: object
          additionalProperties: true
          description: Raw JSON from MCP-Router register endpoint

    ServerRegistryMap:
      type: object
      additionalProperties:
        $ref: "#/components/schemas/ServerRegistryEntry"

    ServerRegistryEntry:
      type: object
      properties:
        transport:
          type: string
          enum: [stdio, sse]
        name:
          type: string
        description:
          type: string
        command:
          type: string
          description: Present when transport is stdio
        args:
          type: array
          items:
            type: string
        url:
          type: string
          description: Present when transport is sse

    RemoveServerResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
        removed:
          type: string

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
        router:
          oneOf:
            - type: object
              additionalProperties: true
            - type: object
              properties:
                error:
                  type: string
        registered_servers:
          type: integer
          minimum: 0

    OkStatus:
      type: object
      properties:
        status:
          type: string
          example: ok

    HTTPError:
      type: object
      properties:
        detail:
          oneOf:
            - type: string
            - type: array
              items: {}